pyproject.toml
[project]
name = "testudo-smart-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "anthropic>=0.49",
    "httpx>=0.28",
]
